+++ legend-engine-language-pure-compiler/src/main/java/org/finos/legend/engine/language/pure/compiler/toPureGraph/ProcessingContext.java
@@ -27,6 +27,7 @@ public class ProcessingContext
 {
     public MutableList<MutableMap<String, ValueSpecification>> inferredVariableList = FastList.newList();
     public MilestoningDatePropagationContext milestoningDatePropagationContext = new MilestoningDatePropagationContext();
+    public boolean isDatePropagationSupported = true;
     private final Stack<String> tags = new Stack<>();
 
     public ProcessingContext(String firstTag)
