+++ legend-engine-language-pure-compiler/src/main/java/org/finos/legend/engine/language/pure/compiler/toPureGraph/ValueSpecificationBuilder.java
@@ -488,15 +488,7 @@ public class ValueSpecificationBuilder implements ValueSpecificationVisitor<org.
     public ValueSpecification visit(AppliedFunction appliedFunction)
     {
         processingContext.push("Applying " + appliedFunction.function);
-        if ((appliedFunction.function.equals("map") || appliedFunction.function.equals("exists")) && appliedFunction.parameters.size() != 0)
-        {
-            ValueSpecification parameterValue = appliedFunction.parameters.get(0).accept(new ValueSpecificationBuilder(this.context, openVariables, processingContext));
-            List<MilestoningStereotype> milestoningStereotype = Milestoning.temporalStereotypes(parameterValue._genericType()._rawType()._stereotypes());
-                if (!milestoningStereotype.isEmpty())
-                {
-                    MilestoningDatePropagationHelper.updateMilestoningPropagationContext((SimpleFunctionExpression) parameterValue, processingContext);
-                }
-        }
+        MilestoningDatePropagationHelper.checkForValidSource(appliedFunction, processingContext);
         if (appliedFunction.function.equals("letFunction"))
         {
             MutableList<org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.valuespecification.ValueSpecification> vs = ListIterate.collect(appliedFunction.parameters, expression -> expression.accept(new ValueSpecificationBuilder(this.context, openVariables, processingContext)));
@@ -512,14 +504,9 @@ public class ValueSpecificationBuilder implements ValueSpecificationVisitor<org.
         Assert.assertTrue(func.getOne() != null, () -> "Can't find a match for function '" + appliedFunction.function + "(" + (func.getTwo() == null ? "?" : LazyIterate.collect(func.getTwo(), v -> (v._genericType() == null ? "?" : v._genericType()._rawType()._name()) + org.finos.legend.pure.m3.navigation.multiplicity.Multiplicity.print(v._multiplicity())).makeString(",")) + ")'", appliedFunction.sourceInformation, EngineErrorType.COMPILATION);
         ValueSpecification result = func.getOne();
         result.setSourceInformation(SourceInformationHelper.toM3SourceInformation(appliedFunction.sourceInformation));
-        
-        if (result instanceof SimpleFunctionExpression && MilestoningDatePropagationHelper.checkGetAllFunctionWithMilestoningContext((SimpleFunctionExpression) result)) {
-            MilestoningDatePropagationHelper.setMilestoningPropagationContext((SimpleFunctionExpression) result, processingContext);
-        }
-        if (result instanceof SimpleFunctionExpression && MilestoningDatePropagationHelper.checkForFilter((SimpleFunctionExpression) result))
-        {
-            MilestoningDatePropagationHelper.updateMilestoningPropagationContextForFilter((SimpleFunctionExpression) result, processingContext);
-        }
+
+        MilestoningDatePropagationHelper.checkForValidSource(appliedFunction, processingContext);
+        MilestoningDatePropagationHelper.updateMilestoningContextFromValidSources(result, processingContext);
         return result;
     }
